requirements.txt
numpy
pandas
matplotlib
numerapi==2.20.0
python-dotenv
pyarrow